+++ packages/amplify-provider-awscloudformation/src/resource-package/resource-packager.ts
@@ -195,6 +195,14 @@ export abstract class ResourcePackager {
     );
   }
 
+  protected resourcesHasApiGatewaysButNotAdminQueries(packagedResources: PackagedResourceDefinition[]): boolean {
+    const { API_CATEGORY, } = Constants;
+    const resources = packagedResources.filter(r => r.resourceName !== 'AdminQueries');
+    return (
+      this.resourcesHasCategoryService(resources, API_CATEGORY.NAME, API_CATEGORY.SERVICE.API_GATEWAY)
+    );
+  }
+
   /**
    * Saves providerMetadata s3 information
    * @param packagedResource
@@ -219,7 +227,7 @@ export abstract class ResourcePackager {
    * @param resources
    */
   protected async generateCategoryCloudFormation(resources: UploadedResourceDefinition[] | PackagedResourceDefinition[]) {
-    if (this.resourcesHasContainers(resources)) {
+    if (this.resourcesHasApiGatewaysButNotAdminQueries(resources)) {
       const { PROVIDER, PROVIDER_NAME } = Constants;
       const { StackName: stackName } = this.amplifyMeta[PROVIDER][PROVIDER_NAME];
       consolidateApiGatewayPolicies(this.context, stackName);