+++ packages/amplify-category-api/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-api@3.3.1...amplify-category-api@3.3.2) (2021-11-19)
+
+**Note:** Version bump only for package amplify-category-api
+
+
+
+
+
 ## [3.3.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-api@3.3.0...amplify-category-api@3.3.1) (2021-11-17)